roadmap/3.8: add schedule, requirements and reformat layout
Adding the current schedule and a description of the requirements for
all the features.

The previous format (a HUGE table) was not workable, so I changed the
layout to use paragraphs and lists instead. It makes the page a little
longer, but at least anyone should be able to add more details that are
currently missing.

Signed-off-by: Niels de Vos <ndevos@redhat.com>